Revert "Merge pull request #428 from MDSplus/dev-gm"
This reverts commit 89aec2584d145c4713a39290151f9796a44e02b7, reversing
changes made to ddeb6370decfed4a831ebce6444533852e002147.

+++ javascope/jScope/MultiWaveform.java
@@ -479,7 +479,7 @@ public class MultiWaveform
         change_limits = true;
     }
 
-    public void setXlimits(double xmin, double xmax)
+    public void setXlimits(float xmin, float xmax)
     {
         if (signals == null)
             return;
